Just comitting a bunch of changes, to keep up to date.
Got some pictures of course. Got a good amount of pages done. Wish there was an easier way to do this, in fact I was going to do this with most of my CYOA books, but seeing how long this will take, I'm just doing this one, and hoping it turns out good.